+++ CHANGELOG.md
@@ -2,5 +2,27 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.4.1 - 2015-08-19
+## [0.5.0] - 2015-10-31
+### Added
+- Gulp
+- CodeClimate reporting and badge
+- runSequence for Gulp
+- jscs for code style
+- jshint for linting
+- dependency badges
+### Changed
+- Gulp, rather than Karma, handles Coveralls reporting
+### Fixed
+- Coveralls reporting and badge
+### Security
+- Adds requireSafe security check to Gulp
+### Updated
+- Updates package.json
+- Style updates in `angular-model.js`
+
+## [0.4.1] - 2015-08-19
 * Adds a CHANGELOG
+
+[Unreleased]: https://github.com/radify/angular-model/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/radify/angular-model/compare/v0.4.1...v0.5.0
+[0.4.1]: https://github.com/radify/angular-model/compare/v0.4.0...v0.4.1